+++ Slicer/Utils/URLResources.cs
@@ -10,7 +10,7 @@ namespace Slicer.Utils
     {
         public static string Project { get { return "/project/"; } }
         public static string Field { get { return "/field/"; } }
-        public static string Index { get { return "/index/"; } }
+        public static string Insert { get { return "/insert/"; } }
         public static string QueryCountEntity { get { return "/query/count/entity/"; } }
         public static string QueryCountEvent { get { return "/query/count/event/"; } }
         public static string QueryCountEntityTotal { get { return "/query/count/entity/total/"; } }